fix: 修复被动引用媒体覆盖
