+++ qq-maid-gateway-rs/src/gateway/event/quoted_payload.rs
@@ -26,7 +26,8 @@ use super::{
 /// 应在群聊 inbound 完成 @机器人/唤醒词/分隔符剥离后、RefIndex enrich 前调用，
 /// 确保检测用的当前正文与最终进入 Core 的正文一致。
 ///
-/// RefIndex 命中时会用索引原文覆盖 `input_parts`，因此本函数只影响 RefIndex miss 的最终状态。
+/// 完整 RefIndex 命中时会用索引原文覆盖 `input_parts`；被动观察命中时会用索引正文
+/// 与这里保留下来的事件媒体合并，因此本检测也负责避免污染文字进入该合并结果。
 ///
 /// 仅在引用上下文恰好只有一个非空 Text part 时启用检测——
 /// 这是 QQ msg_elements 引用消息中最常见的混合形态。多段落或零文字不触发。
